requirements.txt
Flask
Flask-Login
Flask-Dance
Flask-SQLAlchemy
SQLAlchemy
python-dotenv
openai
google-generativeai
langchain
langchain-google-genai
pandas
faiss-cpu
google-cloud-texttospeech
google-cloud-translate
pygame
streamlit
sentence-transformers
googletrans==4.0.0-rc1
